Only deleteItem if file exists at path.
Implemented here partly to ease merge with proposed changes in other pull requests.

+++ FBSnapshotsViewer/Extensions/FileManager+Move.swift
@@ -10,11 +10,13 @@ import Foundation
 
 extension FileManager {
     func moveItem(at fromURL: URL, to toURL: URL) throws {
-        try self.removeItem(at: toURL)
+        try deleteItem(at: toURL)
         try self.copyItem(at: fromURL, to: toURL)
     }
 
     func deleteItem(at url: URL) throws {
-        try self.removeItem(at: url)
+        if fileExists(atPath: url.path) {
+            try self.removeItem(at: url)
+        }
     }
 }
